+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Holidays for Australia
 - Holidays for Bulgaria
+- Holidays for Ukraine
 
 ## [0.9.0] - 2024-06-01